requirements.txt
opencv-python-headless>=4.8.0
numpy>=1.24.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
flask>=2.3.0
python-telegram-bot>=20.6
python-dotenv>=1.0.0
google-generativeai>=0.3.0
